networth/requirements.txt
Django==2.1
djangorestframework==3.9.2
pytz==2019.1
